+++ tests/unit_tests/conftest.py
@@ -1,21 +1,14 @@
 """pytest fixtures."""
 from pathlib import Path
-from typing import List
+from typing import Dict, List
 
 import pytest
-from _pytest.fixtures import FixtureRequest
-from _pytest.tmpdir import TempdirFactory
 from bs4 import BeautifulSoup, element
 from sphinx.testing.path import path
 from sphinx.testing.util import SphinxTestApp
+from TexSoup import TexNode, TexSoup
 
 pytest_plugins = "sphinx.testing.fixtures"  # pylint: disable=invalid-name
-
-
-@pytest.fixture(scope="session")
-def sphinx_test_tempdir(request: FixtureRequest, tmpdir_factory: TempdirFactory) -> path:
-    """Used by sphinx.testing, return the temporary working directory used by Sphinx."""
-    return path(tmpdir_factory.mktemp(request.session.name)).abspath()
 
 
 @pytest.fixture(scope="session")
@@ -29,6 +22,18 @@ def _sphinx_app(app: SphinxTestApp) -> SphinxTestApp:
     """Instantiate a new Sphinx app per test function."""
     app.build()
     yield app
+
+
+@pytest.fixture()
+def ls_out_files(sphinx_app: SphinxTestApp) -> Dict[str, int]:
+    """Return file names and sizes in output directory."""
+    results = {}
+    outdir = sphinx_app.outdir
+    for name in outdir.listdir():
+        file_ = outdir / name
+        if file_.isfile():
+            results[name] = file_.stat().st_size
+    return results
 
 
 @pytest.fixture(name="index_html")
@@ -48,3 +53,19 @@ def img_tags(index_html: BeautifulSoup) -> List[element.Tag]:
 def blockquote_tags(index_html: BeautifulSoup) -> List[element.Tag]:
     """Return all <blockquote> tags in index.html."""
     return index_html.find_all("blockquote")
+
+
+@pytest.fixture(name="python_tex")
+def _python_tex(sphinx_app: SphinxTestApp) -> TexSoup:
+    """Read and parse generated test LaTeX python.tex."""
+    try:
+        text = (Path(sphinx_app.outdir) / "python.tex").read_text(encoding="utf8")
+    except FileNotFoundError:
+        text = (Path(sphinx_app.outdir) / "Python.tex").read_text(encoding="utf8")
+    return TexSoup(text)
+
+
+@pytest.fixture()
+def latex_graphics(python_tex: TexSoup) -> List[TexNode]:
+    """Return all sphinxincludegraphics in python.tex."""
+    return python_tex.find_all("sphinxincludegraphics")